Refactor UniswapFeeService and add package provider
- rename to UniswapFeeProvider
- return Single from getSwapInfo()
- Change UniswapFeeProvider's fee priority to Medium,
